requirements.txt
crewai[tools]
langchain-google-genai
python-dotenv
fastapi
uvicorn
pydantic
google-search-results
